ReportCard: display full calendar for 30 days
The ReportCard uses the past 30 days of data to render an "at a glance" view of diabetic control. Previously, it would filter from the first Monday it detected to the present day, rendering only these days.

Now, the full 30 day calendar is always rendered. The data for each day is retrieved and rendered, and if there is no data, a placeholder colour is used for the PGS blip.