2nd pass at computing basals.
Computing the basal series is now functional. The approach taken here isn't as efficient as the 1st pass - it computes the basal at a 5 minute interval (as in Nightscout), based on either the current treatment or the profile default. This runs quickly enough, and is actually comparably better UX (since basals lasting less than 5mins are unlikely to be relevant for most setups).