refactor: remove old out-of-range check
out-of-range events were likely caused by an incorrect backend API. The backend used to use the timestamp field for querying date ranges, before I discovered that depending on implementation, the timestamp string would not always be in the UTC timezone (+0 offset). So for Brisbane, UTC+10, the API had been returning 10h of data from yesterday. This has been fixed, as we now use the created_at field.