Temp IOB Viz: display delta of temp and profile rate
This is slightly more informative than the temporary basals. The delta of the profile rate and the temporary basal is the corrective dose that the Loop system is issuing. Rather than always displaying the entire dose, we are going to display the delta from the normal basal that is issued. This should make diagnosing incorrect ratios (excluding the basal) much easier.